feat(header): add simple scroll hook for header background transition
