+++ src/components/Header/hooks/useHeaderScroll.hook.ts
@@ -1,0 +1,16 @@
+import { useEffect, useState } from 'react';
+
+export const useHeaderScroll = (threshold = 20) => {
+  const [isScrolled, setIsScrolled] = useState(false);
+
+  useEffect(() => {
+    const handleScroll = () => {
+      setIsScrolled(window.scrollY > threshold);
+    };
+
+    window.addEventListener('scroll', handleScroll);
+    return () => window.removeEventListener('scroll', handleScroll);
+  }, [threshold]);
+
+  return isScrolled;
+};
