--- src/components/Header/hooks/useScrollEffect.hook.ts
@@ -1,16 +1,0 @@
-import { useEffect, useState } from "react";
-
-export const useScrollEffect = (threshold = 20) => {
-  const [isScrolled, setIsScrolled] = useState(false);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      setIsScrolled(window.scrollY > threshold);
-    };
-
-    window.addEventListener("scroll", handleScroll);
-    return () => window.removeEventListener("scroll", handleScroll);
-  }, [threshold]);
-
-  return isScrolled;
-};
