+++ tailwind.config.js
@@ -10,6 +10,19 @@ export default {
         space: ['Space Grotesk', 'sans-serif'],
         mono: ['JetBrains Mono', 'Consolas', 'Monaco', 'monospace'],
       },
+      colors: {
+        js: {
+          DEFAULT: '#f7df1e',
+          hover: 'rgb(247, 223, 30, 0.9)',
+          light: 'rgb(247, 223, 30, 0.1)',
+          medium: 'rgb(247, 223, 30, 0.2)',
+        },
+        dark: {
+          DEFAULT: '#1a1a1a',
+          light: 'rgb(26, 26, 26, 0.3)',
+          medium: 'rgb(26, 26, 26, 0.5)',
+        }
+      }
     },
   },
   plugins: [],
